pending/bellybook/genesis/templates/backend-python/pyproject.toml
[project]
name = "python1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
]
